move lxml.html dependency to worker
create new shared base_web_collector
